コマンドとインタラクションを分離
